CertificateServiceTest with Moskito + jUnit

+++ src/main/java/com/epam/esm/dao/CertificateTagDao.java
@@ -3,7 +3,7 @@ package com.epam.esm.dao;
 
 public interface CertificateTagDao {
 
-    public long createNewCertificateTagRelation(long certificateId, long tagId);
+    public Integer createNewCertificateTagRelation(long certificateId, long tagId);
